refactor(service): split into module folder and extract helper functions
